calendar working, no connetion feature, missing rotation and number of remainders update

+++ app/src/main/java/com/example/franciscoandrade/calendarmobile/model/Day.java
@@ -4,16 +4,16 @@ import java.util.List;
 
 public class Day {
     
-        private short dayNumber;
+        private int dayNumber;
         private DayDetails dayDetailsList;
 
 
 
-        public short getDayNumber() {
+        public int getDayNumber() {
             return dayNumber;
         }
 
-        public void setDayNumber(short dayNumber) {
+        public void setDayNumber(int dayNumber) {
             this.dayNumber = dayNumber;
         }
